requirements.txt
Django>=5.0.2
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
django-allauth>=0.61.1
Pillow>=10.2.0
django-environ==0.11.2
django-filter==23.3
django-widget-tweaks>=1.5.0
python-decouple>=3.8
whitenoise>=6.6.0
gunicorn>=21.2.0
psycopg2-binary==2.9.9
django-storages==1.14.2
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
python-dotenv==1.0.0
stripe==7.6.0
drf-spectacular>=0.27.0
sentry-sdk>=1.39.1
django-redis>=5.4.0 